# public/swagger/swagger.yaml
swagger: '2.0'
info:
  description: LPT APIs
  version: 1.0.0
  title: Frontend APIs
basePath: /api/
...
paths:
...
  /AddRole:
    post:
      tags:
        - Role Management
      summary: Add Role
      description: Add a new role.
      operationId: addRole
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: roleData
          description: Role data including role name
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
      responses:
        '200':
          description: Role added successfully
        '422':
          description: Unprocessable Entity - Validation Error
  /AddPermission:
    post:
      tags:
        - Role Management
      summary: Add Permission
      description: Add a new permission.
      operationId: addPermission
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: permissionData
          description: Permission data including permission name
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
      responses:
        '200':
          description: Permission added successfully
        '422':
          description: Unprocessable Entity - Validation Error
  /AssignRoleToPermission:
    post:
      tags:
        - Role Management
      summary: Assign Role to Permission
      description: Assign a role to a permission.
      operationId: assignRoleToPermission
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: rolePermissionData
          description: Role and permission data
          required: true
          schema:
            type: object
            properties:
              roleName:
                type: string
              permissions:
                type: array
                items:
                  type: string
      responses:
        '200':
          description: Role assigned to permissions successfully
        '404':
          description: Not Found - Role or permission not found
        '422':
          description: Unprocessable Entity - Validation Error
...